`$value` in `slug_update_spaceship` not set if it evalues to false
If you check for `! $value` in the validation in the example callback method `slug_update_spaceship`, a string value of `'0'` (or anything else that PHP evaluates to false) will be ignored and that post meta field will not be updated.

Since `'0'` might be a desired post meta value, I propose changing the validation on this line to instead check `isset` and ensure the string's length is >0.